created template for member dashboard and related tabs | shifted admin dashboard to adminApp
