+++ together_culture/adminApp/templates/Dashboard.html
@@ -1,0 +1,7 @@
+{% include 'adminNavBar.html' %}
+
+<div class="content">
+        <div class="header">🚀 Dashboard</div>
+        <h1>Welcome to Your Dashboard</h1>
+        <p>Manage your insights, events, members, and memberships efficiently with an intuitive interface.</p>
+</div>
