+++ together_culture/adminApp/templates/insights.html
@@ -1,0 +1,9 @@
+{% include 'adminNavBar.html' %}
+
+{% block content %}
+    <div class="content">
+        <div class="header">📊 Insights</div>
+        <h1>Welcome to Insights</h1>
+        <p>Analyze data and track your progress.</p>
+    </div>
+{% endblock %}
