+++ together_culture/adminApp/urls.py
@@ -4,4 +4,5 @@ from . import views
 
 urlpatterns = [
     path("", views.index, name="index"),
+     path('insights/', views.insights, name='insights'),
 ]
