+++ together_culture/adminApp/views.py
@@ -5,3 +5,6 @@ from django.http import HttpResponse
 
 def index(request):
     return HttpResponse("Hello, world. This is the index for admin.")
+
+def insights(request):
+    return render(request, 'insights.html')
