+++ together_culture/loginRegistrationApp/views.py
@@ -37,13 +37,6 @@ def insertUser(request):
 def dashBoard(request):
     return render(request, 'Dashboard.html')
 
-
-def insights(request):
-    return render(request, 'insights.html')
-
-def events(request):
-    return render(request, 'events.html')
-
 def manage_member(request):
     return render(request, 'manage_member.html')
