+++ together_culture/memberApp/templates/digitalContent.html
@@ -1,0 +1,15 @@
+{% include 'navigationBar.html' %}  
+
+<div class="content">
+    <div class="header">📚 Digital Content</div>
+    <h1>Access Exclusive Content</h1>
+    <p>Unlock premium digital content and resources with your membership.</p>
+    <div class="content">
+        {% for content in digitalContent %}
+        <div class="content">
+            <h2>{{ content.title }}</h2>
+            <p>{{ content.description }}</p>
+            <a href="{{ content.link }}">View Content</a>
+        </div>
+        {% endfor %}
+    </div>
