+++ together_culture/memberApp/templates/events.html
@@ -1,0 +1,18 @@
+{% include 'navigationBar.html' %}  
+
+<!--you can remove below code-->
+<div class="content">
+    <div class="header">📅 Events</div>
+    <h1>Upcoming Events</h1>
+    <p>Check out the latest events and RSVP to save your spot!</p>
+    <div class="events">
+        {% for event in events %}
+        <div class="event">
+            <h2>{{ event.title }}</h2>
+            <p>{{ event.date }}</p>
+            <p>{{ event.location }}</p>
+            <p>{{ event.description }}</p>
+            <button>RSVP</button>
+        </div>
+        {% endfor %}
+    </div>
