+++ together_culture/memberApp/templates/memberDashboard.html
@@ -1,0 +1,9 @@
+{% include 'navigationBar.html' %}
+
+
+<!--you can remove belove code  -->
+<div class="content">
+    <div class="header">🚀 Members Dashboard</div>
+    <h1>Welcome to Your Dashboard</h1>
+    <p>Explore events, track your benefits, and manage your profile seamlessly.</p>
+</div>
