+++ together_culture/memberApp/templates/profile.html
@@ -1,0 +1,7 @@
+{% include 'navigationBar.html' %}
+
+<div class="content">
+    <div class="header">👤 My Profile</div>
+    <h1>Update Your Profile</h1>
+    <p>Member contribution in events </p>
+
