+++ together_culture/memberApp/templates/settings.html
@@ -1,0 +1,7 @@
+{% include 'navigationBar.html' %}      
+
+<div class="content">
+    <div class="header">🔧 Settings</div>
+    <h1>Manage Your Account Settings</h1>
+    <p>Update your account settings and preferences.</p>
+</div>
