+++ together_culture/memberApp/urls.py
@@ -1,7 +1,13 @@
 from django.urls import path
 
-from . import views
+from memberApp import views
 
 urlpatterns = [
     path("", views.index, name="index"),
+    path('memberDashboard/', views.memberdashBoard, name='memberDashboard'),
+    path('events/', views.events, name='events'),
+    path('benefits/', views.benefits, name='benefits'),
+    path('digitalContent/', views.digitalContent, name='digitalContent'),
+    path('profile/', views.profile, name='profile'),
+    path('settings/', views.settings, name='settings'),
 ]
