+++ together_culture/together_culture/urls.py
@@ -22,14 +22,7 @@ urlpatterns = [
     path("", include("nonmemberApp.urls")),
     path("member/", include("memberApp.urls")),
     path("admin/", include("adminApp.urls")),
+    path("loginRegistration/", include("loginRegistrationApp.urls")),
     path('djangoadmin/', admin.site.urls),
-    path('register/', views.register, name='register'),
-    path('insertUser/', views.insertUser, name='insertUser'),
-    path('login/', views.login, name='login'),
-    path('dashboard/', views.dashBoard, name='dashboard'),
-    path('validateUser/', views.validateUser, name='validateUser'),
-    path('insights/', views.insights, name='insights'),
-    path('events/', views.events, name='events'),
-    path('manage-member/', views.manage_member, name='manage_member'),
-    path('membership/', views.membership, name='membership'),
+
 ]
